Restructure the app and cascade eslintrc configuration
