+++ .eslintrc
@@ -1,4 +1,5 @@
 {
+  "root": true,
   "extends": [
     "eslint:recommended",
     "plugin:react/recommended",
@@ -31,8 +32,6 @@
   },
   "env": {
     "es6": true,
-    "node": true,
-    "browser": true,
-    "mocha": true
+    "node": true
   }
 }
