+++ app/main/index.js
@@ -47,7 +47,7 @@ app.on('ready', async () => {
     show: false,
   });
 
-  mainWindow.loadFile(path.join(__dirname, 'index.html'));
+  mainWindow.loadFile(path.resolve(path.join(__dirname, '../renderer/index.html')));
 
   // show window once on first load
   mainWindow.webContents.once('did-finish-load', () => {
