+++ app/renderer/app.js
