+++ app/renderer/components/LoggedIn.js
