+++ app/renderer/containers/LoggedInPage.js
