+++ app/renderer/reducers/user.js
