+++ app/renderer/store.js
