+++ test/actions/user.spec.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
-import actions from '../../app/actions/user';
+import actions from '../../app/renderer/actions/user';
 
 const mockStore = configureMockStore([thunk]);
